Add conditional logic to migration

+++ db/migrate/20160725202047_remove_expense_account_id_from_expenses.rb
@@ -1,5 +1,7 @@
 class RemoveExpenseAccountIdFromExpenses < ActiveRecord::Migration
   def change
-    remove_column :expenses, :expense_account_id
+    if column_exist?(:expense, :expense_account_id)
+      remove_column :expenses, :expense_account_id
+    end
   end
 end
